# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LogB",
    "version": "0.1.0",
    "description": "The HTTP API behind LogB, served under `/api`.\n\nTwo credentials work everywhere except issuing and listing tokens: the `logb_session` cookie a browser gets from `POST /auth/login`, and an `Authorization: Bearer logb_pat_...` token for clients that are not browsers. A client obtains its first token by logging in interactively once and calling `POST /auth/tokens`; issuing and listing tokens deliberately refuse bearer credentials, so a leaked token cannot mint replacements. `DELETE /auth/tokens/{id}` is the one exception: a bearer token may revoke itself, so an app can sign out without ever holding a cookie -- but not any other token. `GET /health` lists `token-self-revoke` in its `features` array so a client can detect support before relying on it.\n\n**Writing offline.** `POST /objects/{id}/activities` and `POST /objects/{id}/attachments` accept a `client_op_id`: any string unique to the write, generated once by the client and kept across every retry. The server records it under a unique index, so replaying a write whose response was lost resolves to the row already created instead of duplicating it. Any client that queues writes while offline should use it.\n\nThis document is checked against the running router by `tests/openapi.rs`: a route added, removed or renamed without updating it fails the build."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
  "paths": {
...
    "/types": {
      "get": {
        "summary": "The caller's own object types (built-in types are not listed), sorted by name.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an object type. Send `client_uuid` to make a retry idempotent.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomTypeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomType"
                }
              }
            }
          },
          "200": {
            "description": "Already created: the row this `client_uuid` names",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomType"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name, icon, categories or counter unit, or the name is already used",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`client_uuid` names a row the caller cannot reuse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/types/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }
      ],
      "patch": {
        "summary": "Replace a type's name, icon, categories and counter unit. Its key never changes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomTypeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomType"
                }
              }
            }
          },
          "400": {
            "description": "Invalid field, or the name is already used by another of the caller's types",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such type of the caller's"
          }
        }
      },
      "delete": {
        "summary": "Delete a type. Refused while any non-deleted object uses it.",
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "No such type of the caller's"
          },
          "409": {
            "description": "Still used by objects; `count` says how many",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InUseError"
                }
              }
            }
          }
        }
      }
    },
...
    "/objects/{id}/weight": {
      "get": {
        "summary": "Weight history, newest first",
        "description": "Owned-object measurements ordered by date, created_at and id descending. Deleted entries and dates beyond the reading horizon (tomorrow) are excluded. No attachment or counter queries.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weight measurements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeightPoint"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible object"
          }
        }
      }
    },
    "/objects/{id}/weight/summary": {
      "get": {
        "summary": "Weight summary",
        "description": "Count and aggregate values for an owned object's weight measurements.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weight summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightSummary"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible object"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "logb_session",
        "description": "Set by `POST /auth/login`. HttpOnly, SameSite=Lax, 30 days."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API token from `POST /auth/tokens`. Does not expire; revoke it to end it."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "Category": {
        "type": "string",
        "enum": [
          "maintenance",
          "repair",
          "purchase",
          "inspection",
          "modification",
          "fuel",
          "usage",
          "other",
          "symptom",
          "treatment",
          "appointment",
          "medication",
          "reading",
          "trip",
          "weight",
          "session"
        ],
        "description": "`reading` is an entry that records nothing but a counter value, and requires `counter_value`. `trip` requires `counter_value` and `start_counter`, and is allowed only on an object whose counter_unit is km or mi; its title may be blank."
      },
...
      "CustomType": {
        "type": "object",
        "required": [
          "id",
          "client_uuid",
          "key",
          "name",
          "icon",
          "categories",
          "counter_unit",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "client_uuid": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "description": "`custom:<client_uuid>` -- what an object's `type` holds when it uses this type"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "counter_unit": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "km",
              "mi",
              "h",
              null
            ]
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "CustomTypeInput": {
        "type": "object",
        "required": [
          "name",
          "icon",
          "categories"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "description": "Trimmed; 1-40 characters; unique among the caller's types ignoring case and accents"
          },
          "icon": {
            "type": "string",
            "enum": [
              "document",
              "camera",
              "car",
              "e-bike",
              "bike",
              "motorcycle",
              "home",
              "appliance",
              "tool",
              "body",
              "object"
            ]
          },
          "categories": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Category"
            },
            "description": "Duplicates are dropped, first order kept; `other` is appended when missing"
          },
          "counter_unit": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "km",
              "mi",
              "h",
              null
            ]
          },
          "client_uuid": {
            "type": "string",
            "description": "Create only: an identity minted by the client, stored lower case. A replay answers 200 with the same row."
          }
        }
      },
      "InUseError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "const": "in_use"
          },
          "message": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "description": "How many non-deleted objects still use it"
          }
        }
      },
...
      "WeightPoint": {
        "type": "object",
        "required": [
          "id",
          "date",
          "created_at",
          "weight_grams"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "created_at": {
            "type": "string"
          },
          "weight_grams": {
            "type": "integer"
          }
        }
      },
      "WeightSummary": {
        "type": "object",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer"
          },
          "latest_grams": {
            "type": [
              "integer",
              "null"
            ]
          },
          "previous_grams": {
            "type": [
              "integer",
              "null"
            ]
          },
          "minimum_grams": {
            "type": [
              "integer",
              "null"
            ]
          },
          "maximum_grams": {
            "type": [
              "integer",
              "null"
            ]
          },
          "average_grams": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
...
    }
  }
}